Fix return statement in vmon
