+++ minard/channeldb.py
@@ -253,47 +253,6 @@ def get_nominal_settings(crate, slot, channel):
 
     return dict(zip(keys,row))
 
-def get_most_recent_polling_info(crate, slot, channel):
-    """
-    Returns a dictionary of the cmos and base currents check rates polling 
-    info for a single channel in the detector, for the most recent check rates.
-    """
-    polls = {}
-
-    conn = engine.connect()
-
-    # Get the latest cmos rates
-    result = conn.execute("SELECT * FROM cmos WHERE crate = %s "
-        "AND slot = %s AND channel = %s ORDER BY run DESC LIMIT 1",
-        (crate, slot, channel))
-
-    if result is None:
-        return None, None
-
-    keys = result.keys()
-    row = result.fetchone()
-
-    if row:
-        cmos = dict(zip(keys,row))
-        polls.update(cmos)
-
-    # Get the latest base currents
-    result = conn.execute("SELECT * FROM base WHERE crate = %s "
-        "AND slot = %s AND channel = %s ORDER BY run DESC LIMIT 1",
-        (crate, slot, channel))
-
-    if result is None:
-        return None, None
-
-    keys = result.keys()
-    row = result.fetchone()
-
-    if row:
-        base = dict(zip(keys,row))
-        polls.update(base)
-
-    return polls
-
 def get_discriminator_threshold(crate, slot):
     """
     Get the current discriminator threshold for a specified crate and card.
